# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flow POS SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenant:
    get:
      operationId: getTenant
      tags: [tenant]
      summary: Get current tenant
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenant]
      summary: Update current tenant settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
  /tenant/upgrade-request:
    post:
      operationId: createSubscriptionUpgradeRequest
      tags: [tenant]
      summary: Request a subscription package upgrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpgradeRequestInput"
      responses:
        "200":
          description: Request created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionUpgradeRequest"
        "400":
          description: Invalid input or active pending request already exists

  # ── Subscription Plans ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, name, businessType, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: ["string", "null"]
        businessType:
          type: string
          enum: [fnb, fashion, restaurant, cafe, salon, minimarket]
        status:
          type: string
          enum: [active, suspended, frozen, trial, expired]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        subscriptionPlan:
          type: ["string", "null"]
        subscriptionExpiresAt:
          type: ["string", "null"]
        enableDelivery:
          type: boolean
        deliveryFeeNear:
          type: integer
        deliveryFeeFar:
          type: integer
        showVariants:
          type: boolean
        showToppings:
          type: boolean
        enableCustomerLogin:
          type: boolean
        enableTax:
          type: boolean
        taxPercentage:
          type: number
        enableServiceCharge:
          type: boolean
        serviceChargePercentage:
          type: number
        pointSystemConfig:
          type: object
          properties:
            pointsPerItem:
              type: integer
            minClaimPoints:
              type: integer
            rewardDescription:
              type: string
        defaultCashierName:
          type: ["string", "null"]
        receiptFooter:
          type: ["string", "null"]
        qrisId:
          type: ["string", "null"]
        qrisImageUrl:
          type: ["string", "null"]
        showDeliveryInfo:
          type: boolean
        estimatedDeliveryTime:
          type: ["string", "null"]
        enableOpsHours:
          type: boolean
        opsOpeningTime:
          type: ["string", "null"]
        opsClosingTime:
          type: ["string", "null"]
        createdAt:
          type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        primaryColor:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        bio:
          type: string
        receiptFooter:
          type: string
        defaultCashierName:
          type: string
        qrisId:
          type: string
        qrisImageUrl:
          type: string
        enableDelivery:
          type: boolean
        deliveryFeeNear:
          type: integer
        deliveryFeeFar:
          type: integer
        showVariants:
          type: boolean
        showToppings:
          type: boolean
        enableCustomerLogin:
          type: boolean
        enableTax:
          type: boolean
        taxPercentage:
          type: number
        enableServiceCharge:
          type: boolean
        serviceChargePercentage:
          type: number
        showDeliveryInfo:
          type: boolean
        estimatedDeliveryTime:
          type: string
        enableOpsHours:
          type: boolean
        opsOpeningTime:
          type: string
        opsClosingTime:
          type: string
        pointSystemConfig:
          type: object
          properties:
            pointsPerItem:
              type: integer
            minClaimPoints:
              type: integer
            rewardDescription:
              type: string

    # ── Subscription ──────────────────────────────────────────────────────
...
    SubscriptionUpgradeRequest:
      type: object
      required: [id, tenantId, requestedPlan, billingCycle, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        requestedPlan:
          type: string
          enum: [starter, business, pro, enterprise]
        billingCycle:
          type: string
          enum: [monthly, yearly]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
        updatedAt:
          type: string
    SubscriptionUpgradeRequestInput:
      type: object
      required: [requestedPlan, billingCycle]
      properties:
        requestedPlan:
          type: string
          enum: [starter, business, pro, enterprise]
        billingCycle:
          type: string
          enum: [monthly, yearly]